ScrollToTop component is used to scroll to the top of the page with each route change.
