+++ src/index.js
@@ -6,6 +6,7 @@ import reportWebVitals from './reportWebVitals'
 import { BrowserRouter } from 'react-router-dom'
 import { QueryClient, QueryClientProvider } from 'react-query'
 import {ContextProvider} from './Context'
+import ScrollToTop from './Components/scrollToTop'
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 const queryClient = new QueryClient();
@@ -15,6 +16,7 @@ root.render(
   <BrowserRouter>
     <QueryClientProvider client={queryClient}>
       <ContextProvider>
+        <ScrollToTop />
         <App />
       </ContextProvider>
     </QueryClientProvider>
